Change Type of list to BigDecimal for better precision (reverted from commit 6f907d06b57332d98ddaeaca19db021b3d036632)

+++ src/main/java/pl/dawidbasa/crediAnalyser/Credit/CreditService.java
@@ -1,6 +1,5 @@
 package pl.dawidbasa.crediAnalyser.Credit;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
@@ -10,8 +9,8 @@ public interface CreditService {
 	public void saveCredit(Credit credit);
 	List<Credit> findAll();
 	public List<Credit> sortCreditsByCreditMargin(List<Credit> credits);
-	public List<BigDecimal> calculateAllDecreasingInstalments(Credit credit);
-	public Map<String, BigDecimal> calculateDecrasingInstalmentDetails(Credit credit);
+	public List<Double> calculateAllDecreasingInstalments(Credit credit);
+	public Map<String, Double> calculateDecrasingInstalmentDetails(Credit credit);
 	public Map<String, Double> calculateConstantInstalmentDetails(Credit credit);
 	
 }
